dropout-back-main/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.4.0
pymongo==4.8.0
pydantic<2.0.0
email-validator==2.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
scikit-learn
pandas
numpy
python-dotenv==1.0.0
joblib
seaborn
matplotlib
requests==2.31.0
google-auth==2.33.0
httpx==0.28.1
google-generativeai
PyMuPDF
python-docx